Trac #3: PSP modular build support
Adds necessary build framework to support adding user-defined code modules into the PSP.

Note that this **must** be merged with a corresponding change to CFE that updates the
main CMake build scripts to support this.

The modules are statically linked into the build rather than dynamically loaded as it
would be as a CFS application or via OS_ModuleLoad.   This allows some parts of the PSP
functionality to diverge without forking the entire PSP for a particular board.  An
example use case would be EEPROM/NVRAM - some boards may have no EEPROM and others might
have an EEPROM connected via SPI or I2C and others may have a memory mapped NVRAM.  It
should not be necessary to clone an entire BSP just to handle these differences.

Note that this is only built if using the CMake scripts since these generate the required
intermediate files during the configuration step.