Trac #6: Dependency management
Several changes to pave the way for easier dependency management:

1. The PSP should *NOT* need to include anything from ES.  This was
being done mostly for the reset types, but in fact the reset types
are really defined by the PSP and the ES is using them.  This change
moves the actual reset type and subtype #defines into the cfe_psp.h
file rather than the cfe_es.h file, which removes the need to include
cfe_es.h for reset types.

A global search-and-replace was done across __ALL__ PSP implementations
to change the CFE_ES reset type macros to the new PSP-defined macros.

(Note that a separate change on the CFE side updates CFE ES to use the
PSP definitions and deprecates the CFE_ES definitions of the reset types)

2. The PSP should *NOT* need to include anything from cfe_platform_cfg.h
This was done mostly for memory size requirements.  In this case the
new build system offers a way to obtain these sizes at runtime without
directly including the platform config.

Only the "pc-linux" PSP has been fully updated to take advantage of the
new features here to totally scrub the use of cfe_es.h and
cfe_platform_cfg.h.  The other PSPs continue to operate as before, aside
from the reset types change above.

Fixes #6