Trac #15: Compatibility with CFE "const" API
On the CFE side, trac 28 modifies several API calls to add a "const"
designation to the pointer.  This is important for many reasons,
but the most important reason to make this change is that the data
can be directly supplied from a structure on the PSP side that is
also "const".

On some architectures, it is possible (via a linker script) to
actually keep the read-only data section of an executable in
ROM rather than copying it to RAM.

On many platforms, writing to a "char *" string literal will
actually crash the process (segfault), so it is very important
that the pointer is qualified as a "const char *" so the compiler
will catch any writes to it.