Trac #23: Integrate pc-linux PSP with timebase API
Configure the pc-linux PSP to setup a "master" timebase
object with which the CFE_TIME task can set up its own 1Hz
callback timer.  This makes things cleaner and also avoids
a race condition where the timer callback could be invoked
before the global variables and semaphores are initialized.

As a fallback, if this fails, it sets up the same signal
based callback as was done previously.  This preserves
compatibility with old OSAL libraries that do not implement
this feature.