Trac #3: EEPROM implementations as PSP modules
Proof-of-concept for loadable PSP modules.  This removes the "fake eeprom" support
that was in the pc-linux PSP and turns it into a module called "eeprom_mmap_file".

This also adds a second module that stubs out the same API calls that can be used
on PSPs that do not support or emulate an EEPROM at all.