Trac #81: Update pc-rtems for CFE 6.6 + RTEMS 4.11
The pc-rtems PSP was originally developed using the RTEMS development
branch prior to the 4.11 release.  Since the RTEMS 4.11 official
release some functions are now marked as deprecated.

This commit contains all changes required to use the pc-rtems PSP
with RTEMS 4.11.2 (latest official release at this time).

Some aspects were streamlined for ease of use, mainly that GRUB
is not required to boot the image in QEMU.  The README.txt file
is also updated to reflect these changes and the commands and
sequence in this file has been re-validated.

Additionally, this contains a RTEMS.cmake module which can be
directly included in the build rather than requiring the user
to patch their cmake installation.