Trac #73: Update MCP750 FPU exception flags
From the bug description:

It was found the CFE_PSP_SetDefaultExceptionEnvironment function
defined in cfe_psp_exception.c does not set all the required flags
in order to produce a task restart following a CPU or Float Point
exception produced via task running under 6.9. The Recoverable
Interrupt flag (_PPC_MSR_RI) is missing in the call to vxMsrSet.

It was recommended by WindRiver to update how we setup the default
exception environment using the vxMsrGet call.

Tested this with a vxworks 6.9 toolchain and it builds.