Trac #99: Fix CFE_PSP_Main() prototype
This is the entry point function from the BSP and it was defined
as taking two parameters, but all PSP's that implemented this function
ignored the two parameters and used values supplied through
other means instead.

This is just a clean-up to make CFE_PSP_Main() take no arguments,
which was the reality already.