Trac #104: mcp750-vxworks PSP CMake build
Adds a VxWorks-CFE platform module with the required
CMake logic to build code for the VxWorks-6.9 platform.

Minor updates to the mcp750-vxworks PSP to build the code
and call the entry point in a manner similar to the way
that pc-linux and pc-rtems do this, using the entry point
specified in the target_config object.